refactor(stdlib): remove hardcoded type checking for stdlib modules
- Remove hardcoded log/json/math handling from expr_call.go
- Stdlib modules now use proper @extern wrapper functions
- Create log.desi stdlib module with info/debug/warn/error
- Add missing json.desi wrapper functions: is_int, get_int, get_float
- Fix json.c __json_get_int to return int instead of int64_t
- Fix emit_call.go __json_get_int declaration to use i32
- Add stdlib import validation to catch use without import
- Fix types.md docs: change section header to lowercase 'none'

All 282 tests pass. Import patterns (import, from import, from import *)
all work correctly with the new stdlib architecture.

+++ compiler/runtime/json.c
@@ -495,9 +495,9 @@ int __json_is_int(JsonNode* node) {
 }
 
 // Get number as integer (Rust-style explicit accessor)
-int64_t __json_get_int(JsonNode* node) {
+int __json_get_int(JsonNode* node) {
     if (!node || node->type != JSON_NUMBER) return 0;
-    return (int64_t)node->num_val;
+    return (int)node->num_val;
 }
 
 // Rename get_number to get_float for clarity (alias)
